refactor: remove extra view
