+++ apps/admin-ui/src/spa/my-units/[id]/UnitReservations.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import { breakpoints } from "common/src/common/style";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
@@ -7,7 +7,17 @@ import Legend from "@/component/reservations/requested/Legend";
 import { legend } from "./eventStyleGetter";
 import { UnitCalendar } from "./UnitCalendar";
 import { useUnitResources } from "./hooks";
-import { fromUIDate, isValidDate } from "common/src/common/util";
+import { fromUIDate, isValidDate, toUIDate } from "common/src/common/util";
+import { startOfDay } from "date-fns";
+import { Button } from "hds-react";
+import { useNavigate, useParams, useSearchParams } from "react-router-dom";
+import { useLocation } from "react-use";
+import { AutoGrid, VerticalFlex } from "@/styles/layout";
+import { useReservationUnitTypes } from "@/hooks";
+import DayNavigation from "./DayNavigation";
+import { HR } from "@/component/Table";
+import { SearchTags } from "@/component/SearchTags";
+import { MultiSelectFilter } from "@/component/QueryParamFilters";
 
 type Props = {
   // date in ui string format
@@ -33,7 +43,7 @@ const LegendContainer = styled.div`
   }
 `;
 
-export function UnitReservations({
+function UnitReservationsInner({
   begin,
   unitPk,
   reservationUnitTypes,
@@ -72,5 +82,109 @@ export function UnitReservations({
         </Legends>
       </LegendContainer>
     </>
+  );
+}
+
+const HorisontalFlexWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+`;
+
+type Params = {
+  unitId: string;
+  reservationUnitId: string;
+};
+
+export function UnitReservations(): JSX.Element {
+  const { hash } = useLocation();
+  const [queryParams] = useSearchParams();
+
+  // date in UI string format
+  const queryParamsDate = queryParams.get("date");
+  const date = queryParamsDate != null ? fromUIDate(queryParamsDate) : null;
+  const initialDate =
+    queryParamsDate != null && date && isValidDate(date)
+      ? queryParamsDate
+      : (toUIDate(startOfDay(new Date())) ?? "");
+  const [begin, setBegin] = useState<string>(initialDate);
+  const { unitId } = useParams<Params>();
+  const { t } = useTranslation();
+  const history = useNavigate();
+
+  const onDateChange = (dateString: string) => {
+    setBegin(dateString);
+
+    // TODO should use setQueryParams for consistency
+    // TODO there should not be state duplication either query params or useState, not both
+    // TODO this should not be validated, it should be validated on use
+    // because modifying the query string will will crash the frontend
+    const newDate = fromUIDate(dateString);
+    if (newDate && isValidDate(newDate)) {
+      history({
+        hash,
+        search: `?date=${dateString}`,
+      });
+    }
+  };
+
+  const { options: reservationUnitTypeOptions } = useReservationUnitTypes();
+
+  const translateTag = (key: string, value: string) => {
+    switch (key) {
+      case "reservationUnitType":
+        return (
+          reservationUnitTypeOptions.find((u) => u.value === Number(value))
+            ?.label ?? ""
+        );
+      default:
+        return "";
+    }
+  };
+
+  const [searchParams] = useSearchParams();
+  const reservationUnitTypes = searchParams
+    .getAll("reservationUnitType")
+    .map(Number)
+    .filter(Number.isInteger);
+
+  return (
+    <VerticalFlex>
+      <AutoGrid>
+        <MultiSelectFilter
+          style={{
+            zIndex: "var(--tilavaraus-admin-stack-select-over-calendar)",
+          }}
+          name="reservationUnitType"
+          options={reservationUnitTypeOptions}
+        />
+      </AutoGrid>
+      <SearchTags hide={["date"]} translateTag={translateTag} />
+      <HR />
+      <HorisontalFlexWrapper>
+        <Button
+          variant="secondary"
+          theme="black"
+          size="small"
+          onClick={() => {
+            onDateChange(toUIDate(startOfDay(new Date())));
+          }}
+        >
+          {t("common.today")}
+        </Button>
+        <DayNavigation date={begin} onDateChange={onDateChange} />
+        <div />
+      </HorisontalFlexWrapper>
+      {/* TODO missing unitId is an error, not return null */}
+      {unitId ? (
+        <UnitReservationsInner
+          reservationUnitTypes={reservationUnitTypes}
+          unitPk={unitId}
+          key={begin}
+          begin={begin}
+        />
+      ) : null}
+    </VerticalFlex>
   );
 }
